feat(integrations): migrate comms adapters — email, teams (P2)
Completes the comms category. Both are driver/adapter-injected services that
exercise the generalized `ctx.config` path (no HTTP auth):

- email — SMTP `EmailTransport` + IMAP `ImapClient` adapters via ctx.config;
  email_send keeps requiresConfirmation. Shim preserves the construction-time
  ConfigError throws and conditional tool list.
- teams — Incoming Webhook (ctx.fetch + ctx.config) and Bot Framework client
  (ctx.config) paths; adaptiveCard/messageCard builders + card types live in
  the service and are re-exported by the deprecated tools shim.

Both packages green: tools' existing tests pass unchanged (284), integrations
coverage 95.37%, all 9 quality gates pass.
