+++ packages/integrations/src/services/index.ts
@@ -4,3 +4,5 @@ export { slackIntegration } from './slack'
 export { discordIntegration } from './discord'
 export { gmailIntegration } from './gmail'
 export { twilioIntegration } from './twilio'
+export * from './email'
+export * from './teams'
