+++ packages/tools/src/integrations/teams.ts
@@ -1,213 +1,60 @@
-import { ConfigError, ErrorCodes, ToolError, defineTool } from '@agentskit/core'
+import { ConfigError, ErrorCodes, type ToolDefinition } from '@agentskit/core'
+import {
+  teamsIntegration,
+  toToolDefinitions,
+  type TeamsBotClient,
+} from '@agentskit/integrations'
 import type { HttpToolOptions } from './http'
 
-/**
- * Microsoft Teams integration. Two outbound paths:
- *
- *  - Incoming Webhook (no app registration required): pass `webhookUrl`.
- *    Best for one-way notifications, supports MessageCard + Adaptive Cards.
- *
- *  - Bot Framework (rich bidirectional): pass a `TeamsBotClient` adapter.
- *    botbuilder is not bundled — wrap it (or your Graph/REST client) to
- *    keep this package driver-free. Auth (app secret, certificate,
- *    managed identity) is handled inside your adapter.
- *
- * Inbound activity routing (`message`, `mentioned`, etc.) lives in
- * `@agentskit/triggers` (AgentsKitOS) — this package only owns tool
- * primitives, not long-running listeners.
- */
+export { adaptiveCard, messageCard } from '@agentskit/integrations'
+export type {
+  TeamsAdaptiveCardAction,
+  TeamsAdaptiveCard,
+  TeamsMessageCard,
+  TeamsBotMessage,
+  TeamsBotSendResult,
+  TeamsBotClient,
+} from '@agentskit/integrations'
 
+/** @deprecated Moved to `@agentskit/integrations` (services/teams). */
 export interface TeamsWebhookConfig extends HttpToolOptions {
   webhookUrl: string
 }
 
-export interface TeamsBotMessage {
-  /** Conversation id (channel or 1:1). */
-  conversationId: string
-  serviceUrl?: string
-  text?: string
-  /** Optional Adaptive Card or MessageCard attachment. */
-  card?: TeamsAdaptiveCard | TeamsMessageCard
-  /** Reply to a specific activity id (creates a thread reply). */
-  replyToId?: string
-}
-
-export interface TeamsBotSendResult {
-  id: string
-  conversationId: string
-}
-
-export interface TeamsBotClient {
-  send: (msg: TeamsBotMessage) => Promise<TeamsBotSendResult>
-}
-
+/** @deprecated */
 export interface TeamsBotConfig {
   client: TeamsBotClient
 }
 
+/** @deprecated */
 export interface TeamsConfig {
   webhook?: TeamsWebhookConfig
   bot?: TeamsBotConfig
 }
 
-export interface TeamsAdaptiveCardAction {
-  type: 'Action.OpenUrl' | 'Action.Submit'
-  title: string
-  url?: string
-  data?: Record<string, unknown>
+function webhookTool(config: TeamsWebhookConfig): ToolDefinition {
+  return toToolDefinitions(teamsIntegration, {
+    config: { webhook: { webhookUrl: config.webhookUrl, headers: config.headers, timeoutMs: config.timeoutMs } },
+    fetch: config.fetch,
+  }).find((t) => t.name === 'teams_send_webhook')!
 }
 
-export interface TeamsAdaptiveCard {
-  contentType: 'application/vnd.microsoft.card.adaptive'
-  content: {
-    $schema: 'http://adaptivecards.io/schemas/adaptive-card.json'
-    type: 'AdaptiveCard'
-    version: '1.5'
-    body: Array<Record<string, unknown>>
-    actions?: TeamsAdaptiveCardAction[]
-  }
+function botTool(config: TeamsBotConfig): ToolDefinition {
+  return toToolDefinitions(teamsIntegration, { config: { bot: { client: config.client } } }).find(
+    (t) => t.name === 'teams_send_bot',
+  )!
 }
 
-export interface TeamsMessageCard {
-  contentType: 'application/vnd.microsoft.teams.card.o365connector'
-  content: {
-    '@type': 'MessageCard'
-    '@context': 'https://schema.org/extensions'
-    summary?: string
-    themeColor?: string
-    title?: string
-    text?: string
-  }
-}
-
-/**
- * Build a Teams Adaptive Card payload (v1.5). Keep `body` simple —
- * one TextBlock per paragraph; pass `actions` for buttons.
- */
-export function adaptiveCard(input: {
-  title?: string
-  text?: string
-  facts?: Array<{ title: string; value: string }>
-  actions?: TeamsAdaptiveCardAction[]
-}): TeamsAdaptiveCard {
-  const body: Array<Record<string, unknown>> = []
-  if (input.title) {
-    body.push({ type: 'TextBlock', size: 'Large', weight: 'Bolder', text: input.title, wrap: true })
-  }
-  if (input.text) {
-    body.push({ type: 'TextBlock', text: input.text, wrap: true })
-  }
-  if (input.facts && input.facts.length > 0) {
-    body.push({ type: 'FactSet', facts: input.facts })
-  }
-  return {
-    contentType: 'application/vnd.microsoft.card.adaptive',
-    content: {
-      $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
-      type: 'AdaptiveCard',
-      version: '1.5',
-      body,
-      ...(input.actions && input.actions.length > 0 ? { actions: input.actions } : {}),
-    },
-  }
-}
-
-/**
- * Build a legacy O365 connector MessageCard. Simpler than Adaptive
- * Cards but still rendered by Teams Incoming Webhooks.
- */
-export function messageCard(input: {
-  title?: string
-  text?: string
-  summary?: string
-  themeColor?: string
-}): TeamsMessageCard {
-  return {
-    contentType: 'application/vnd.microsoft.teams.card.o365connector',
-    content: {
-      '@type': 'MessageCard',
-      '@context': 'https://schema.org/extensions',
-      summary: input.summary ?? input.title ?? 'Notification',
-      themeColor: input.themeColor,
-      title: input.title,
-      text: input.text,
-    },
-  }
-}
-
-export function teamsSendWebhook(config: TeamsWebhookConfig) {
+/** @deprecated import from `@agentskit/integrations`. */
+export function teamsSendWebhook(config: TeamsWebhookConfig): ToolDefinition {
   if (!config.webhookUrl) {
-    throw new ConfigError({
-      code: ErrorCodes.AK_CONFIG_INVALID,
-      message: 'teamsSendWebhook: webhookUrl is required',
-    })
+    throw new ConfigError({ code: ErrorCodes.AK_CONFIG_INVALID, message: 'teamsSendWebhook: webhookUrl is required' })
   }
-  const fetchImpl = config.fetch ?? globalThis.fetch
-  const timeoutMs = config.timeoutMs ?? 20_000
-  return defineTool({
-    name: 'teams_send_webhook',
-    description: 'Post a message or Adaptive Card to a Microsoft Teams channel via Incoming Webhook.',
-    schema: {
-      type: 'object',
-      properties: {
-        text: { type: 'string', description: 'Plain text body. Ignored when `card` is provided.' },
-        title: { type: 'string', description: 'Convenience: wrap into a MessageCard with this title.' },
-        card: {
-          type: 'object',
-          description: 'Pre-built Adaptive Card or MessageCard payload (use `adaptiveCard()`/`messageCard()` builders).',
-        },
-      },
-    } as const,
-    async execute(args) {
-      if (!fetchImpl) {
-        throw new ToolError({
-          code: ErrorCodes.AK_TOOL_EXEC_FAILED,
-          message: 'teams_send_webhook: no fetch available',
-          hint: 'Run on Node ≥ 18 or pass config.fetch.',
-        })
-      }
-      const text = typeof args.text === 'string' ? args.text : undefined
-      const title = typeof args.title === 'string' ? args.title : undefined
-      const card = (args.card ?? undefined) as TeamsAdaptiveCard | TeamsMessageCard | undefined
-
-      let payload: Record<string, unknown>
-      if (card) {
-        payload = { type: 'message', attachments: [card] }
-      } else if (text || title) {
-        payload = { type: 'message', attachments: [messageCard({ title, text })] }
-      } else {
-        throw new ToolError({
-          code: ErrorCodes.AK_TOOL_INVALID_INPUT,
-          message: 'teams_send_webhook: provide text, title, or card',
-        })
-      }
-
-      const controller = new AbortController()
-      const timer = setTimeout(() => controller.abort(), timeoutMs)
-      try {
-        const response = await fetchImpl(config.webhookUrl, {
-          method: 'POST',
-          headers: { 'content-type': 'application/json', ...config.headers },
-          body: JSON.stringify(payload),
-          signal: controller.signal,
-        })
-        if (!response.ok) {
-          const body = await response.text().catch(() => '')
-          throw new ToolError({
-            code: ErrorCodes.AK_TOOL_EXEC_FAILED,
-            message: `teams_send_webhook: HTTP ${response.status} ${response.statusText}: ${body.slice(0, 200)}`,
-            hint: 'Verify the webhook URL is current and the channel still exists.',
-          })
-        }
-        return { ok: true, status: response.status }
-      } finally {
-        clearTimeout(timer)
-      }
-    },
-  })
+  return webhookTool(config)
 }
 
-export function teamsSendBot(config: TeamsBotConfig) {
+/** @deprecated import from `@agentskit/integrations`. */
+export function teamsSendBot(config: TeamsBotConfig): ToolDefinition {
   if (!config.client) {
     throw new ConfigError({
       code: ErrorCodes.AK_CONFIG_INVALID,
@@ -215,59 +62,16 @@ export function teamsSendBot(config: TeamsBotConfig) {
       hint: 'Provide a TeamsBotClient adapter (e.g. wrap botbuilder TurnContext.sendActivity).',
     })
   }
-  const client = config.client
-  return defineTool({
-    name: 'teams_send_bot',
-    description: 'Send a message or Adaptive Card via the configured Teams Bot Framework client.',
-    schema: {
-      type: 'object',
-      properties: {
-        conversation_id: { type: 'string' },
-        service_url: { type: 'string' },
-        text: { type: 'string' },
-        card: { type: 'object' },
-        reply_to_id: { type: 'string', description: 'Activity id to reply in-thread.' },
-      },
-      required: ['conversation_id'],
-    } as const,
-    async execute(args) {
-      const text = typeof args.text === 'string' ? args.text : undefined
-      const card = (args.card ?? undefined) as TeamsAdaptiveCard | TeamsMessageCard | undefined
-      if (!text && !card) {
-        throw new ToolError({
-          code: ErrorCodes.AK_TOOL_INVALID_INPUT,
-          message: 'teams_send_bot: provide text or card',
-        })
-      }
-      try {
-        const result = await client.send({
-          conversationId: String(args.conversation_id),
-          serviceUrl: typeof args.service_url === 'string' ? args.service_url : undefined,
-          text,
-          card,
-          replyToId: typeof args.reply_to_id === 'string' ? args.reply_to_id : undefined,
-        })
-        return { id: result.id, conversationId: result.conversationId }
-      } catch (err) {
-        throw new ToolError({
-          code: ErrorCodes.AK_TOOL_EXEC_FAILED,
-          message: `teams_send_bot: ${err instanceof Error ? err.message : String(err)}`,
-          hint: 'Check Bot Framework credentials and that the conversation reference is still valid.',
-        })
-      }
-    },
-  })
+  return botTool(config)
 }
 
-export function teams(config: TeamsConfig) {
-  const tools = []
-  if (config.webhook) tools.push(teamsSendWebhook(config.webhook))
-  if (config.bot) tools.push(teamsSendBot(config.bot))
+/** @deprecated import from `@agentskit/integrations`. */
+export function teams(config: TeamsConfig): ToolDefinition[] {
+  const tools: ToolDefinition[] = []
+  if (config.webhook) tools.push(webhookTool(config.webhook))
+  if (config.bot) tools.push(botTool(config.bot))
   if (tools.length === 0) {
-    throw new ConfigError({
-      code: ErrorCodes.AK_CONFIG_INVALID,
-      message: 'teams: provide at least one of `webhook` or `bot`',
-    })
+    throw new ConfigError({ code: ErrorCodes.AK_CONFIG_INVALID, message: 'teams: provide at least one of `webhook` or `bot`' })
   }
   return tools
 }
